+++ backend/src/main/java/com/funeat/review/dto/RankingReviewsResponse.java
@@ -1,4 +1,4 @@
-package com.funeat.review.presentation.dto;
+package com.funeat.review.dto;
 
 import java.util.List;
