+++ backend/src/main/java/com/funeat/review/dto/ReviewCreateRequest.java
@@ -1,4 +1,4 @@
-package com.funeat.review.presentation.dto;
+package com.funeat.review.dto;
 
 import java.util.List;
 import javax.validation.constraints.NotBlank;
