+++ backend/src/main/java/com/funeat/review/presentation/ReviewApiController.java
@@ -4,10 +4,10 @@ import com.funeat.auth.dto.LoginInfo;
 import com.funeat.auth.util.AuthenticationPrincipal;
 import com.funeat.product.application.ProductService;
 import com.funeat.review.application.ReviewService;
-import com.funeat.review.presentation.dto.RankingReviewsResponse;
-import com.funeat.review.presentation.dto.ReviewCreateRequest;
-import com.funeat.review.presentation.dto.ReviewFavoriteRequest;
-import com.funeat.review.presentation.dto.SortingReviewsResponse;
+import com.funeat.review.dto.RankingReviewsResponse;
+import com.funeat.review.dto.ReviewCreateRequest;
+import com.funeat.review.dto.ReviewFavoriteRequest;
+import com.funeat.review.dto.SortingReviewsResponse;
 import java.net.URI;
 import javax.validation.Valid;
 import org.springframework.data.domain.Pageable;
