+++ backend/src/test/java/com/funeat/acceptance/review/ReviewAcceptanceTest.java
@@ -43,10 +43,10 @@ import com.funeat.member.domain.Member;
 import com.funeat.member.domain.favorite.ReviewFavorite;
 import com.funeat.product.domain.Product;
 import com.funeat.review.domain.Review;
-import com.funeat.review.presentation.dto.RankingReviewDto;
-import com.funeat.review.presentation.dto.ReviewCreateRequest;
-import com.funeat.review.presentation.dto.ReviewFavoriteRequest;
-import com.funeat.review.presentation.dto.SortingReviewDto;
+import com.funeat.review.dto.RankingReviewDto;
+import com.funeat.review.dto.ReviewCreateRequest;
+import com.funeat.review.dto.ReviewFavoriteRequest;
+import com.funeat.review.dto.SortingReviewDto;
 import com.funeat.tag.domain.Tag;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
