+++ backend/src/test/java/com/funeat/review/application/ReviewServiceTest.java
@@ -38,8 +38,8 @@ import com.funeat.member.dto.MemberReviewDto;
 import com.funeat.member.exception.MemberException.MemberNotFoundException;
 import com.funeat.product.exception.ProductException.ProductNotFoundException;
 import com.funeat.review.domain.Review;
+import com.funeat.review.dto.SortingReviewDto;
 import com.funeat.review.exception.ReviewException.ReviewNotFoundException;
-import com.funeat.review.presentation.dto.SortingReviewDto;
 import com.funeat.tag.domain.Tag;
 import java.util.List;
 import java.util.stream.Collectors;
